py/requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-cov>=5.0.0
numpy>=1.26.0
scipy>=1.12.0
